pyproject.toml
[project]
name = "blogapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.2.0",
    "cryptography>=46.0.3",
    "fastapi>=0.127.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
